Conflicts:
	db/database.hpp
	db/node.cpp
	db/node.hpp
	db/trie.cpp
	sched/nodes/serial.hpp
	sched/serial.cpp
	sched/serial_ui.cpp
	vm/state.cpp
	vm/state.hpp

+++ process/remote.hpp
@@ -21,7 +21,7 @@ class remote
 {
 public:
    
-   typedef int remote_id;
+   typedef uint64_t remote_id;
    
    static const remote_id LEADER_RANK = 0;
    static remote *self;
